Split web progress source pipelines
